refactor(datatable): move column helpers to columns/ (refs #298)
