+++ src/components/DataTablePanel.tsx
@@ -25,12 +25,12 @@ import { useApplyTransformation } from 'hooks/useApplyTransformation';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { DatatableOptions } from 'types';
 import { BuildColumnDefs, ConvertDataFrameToDataTableFormat } from 'datatable/dataHelpers';
-import { ApplyColumnWidthHints } from 'datatable/columnWidthHints';
+import { ApplyColumnWidthHints } from 'datatable/columns/columnWidthHints';
 import { buildSearchBarLayout } from 'datatable/layout/buildSearchBarLayout';
 import { datatableThemedStyles } from './styles';
 import { GetDataTransformerID } from 'pipeline/transformations';
 import { DTColumnType } from 'datatable/types';
-import { ApplyColumnAliases } from 'datatable/columnAliasing';
+import { ApplyColumnAliases } from 'datatable/columns/columnAliasing';
 
 interface Props extends PanelProps<DatatableOptions> { }
