+++ src/datatable/columns/columnStyles.ts
@@ -1,5 +1,5 @@
 import { ColumnStyleItemType } from "components/options/columnstyles/types";
-import { DTColumnType } from "./types";
+import { DTColumnType } from "../types";
 
 export const ApplyColumnStyles = (columns: DTColumnType[], columnStyles: ColumnStyleItemType[]) => {
   for (const item of columns) {
